Version 1 complete.

+++ template/.simple-web-client/build.sh
@@ -1,15 +1,31 @@
 #!/bin/bash
 GREEN='\033[0;32m'
+RED='\033[0;31m'
 RESET='\033[0m'
 
-echo -e "\n${GREEN}-> Clearing dist directory${RESET}"
+echo -e "\n${GREEN}-> Running prettier${RESET}"
+if ! npx prettier . --write; then
+    echo -e "\n${RED}-> Failed to run prettier${RESET}"
+    exit 1
+fi
+
+echo -e "\n${GREEN}-> Compiling src${RESET}"
 rm -rf dist
+if ! npx tsc; then
+    echo -e "\n${RED}-> Failed to compile src${RESET}"
+    exit 1
+fi
 
-echo -e "\n${GREEN}-> Running Prettier${RESET}"
-npx prettier . --write
-
-echo -e "\n${GREEN}-> Compiling TypeScript${RESET}"
-npx tsc
+echo -e "\n${GREEN}-> Compiling specs${RESET}"
+rm -rf spec/dist
+if ! npx tsc -p spec; then
+    echo -e "\n${RED}-> Failed to compile specs${RESET}"
+else 
+  echo -e "\n${GREEN}-> Running specs${RESET}"
+  if ! npx jasmine-browser-runner runSpecs; then
+      echo -e "\n${RED}-> Specs failed!${RESET}"
+  fi
+fi
 
 echo -e "\n${GREEN}-> Assembling files${RESET}"
 mkdir dist
@@ -18,10 +34,10 @@ mkdir dist/simple-web-client
 cp -r .simple-web-client/lib/* dist/simple-web-client
 find dist -name "*.ts" -type f -delete
 
-echo -e "\n${GREEN}-> Replacing env${RESET}"
-mv dist/envs/env.build.js dist/envs/env.js
-
-echo -e "\n${GREEN}-> Removing client listener${RESET}"
-sed -i 's;<script type="module" src="/client-listener.js"></script>;;g' dist/index.html
+echo -e "\n${GREEN}-> Replacing env file${RESET}"
+if ! mv dist/envs/env.build.js dist/envs/env.js; then
+    echo -e "\n${RED}-> Failed to replace env file${RESET}"
+    exit 1
+fi
 
 echo -e "\n${GREEN}-> Build complete! Check the dist directory!${RESET}"
